+++ src/main/scala/org/mmarini/scalarl/v2/agents/ExpSarsaAgent.scala
@@ -35,6 +35,7 @@ import com.typesafe.scalalogging.LazyLogging
 import io.circe.ACursor
 import org.deeplearning4j.nn.multilayer.MultiLayerNetwork
 import org.deeplearning4j.util.ModelSerializer
+import org.mmarini.scalarl.v2.Utils._
 import org.mmarini.scalarl.v2._
 import org.nd4j.linalg.activations.Activation
 import org.nd4j.linalg.api.ndarray.INDArray
@@ -67,15 +68,15 @@ import org.nd4j.linalg.ops.transforms.Transforms._
  */
 case class ExpSarsaAgent(net: MultiLayerNetwork,
                          noActions: Int,
-                         avgReward: Double,
-                         epsilon: Double,
-                         kappa: Double,
-                         beta: Double,
+                         avgReward: INDArray,
+                         epsilon: INDArray,
+                         kappa: INDArray,
+                         beta: INDArray,
                          model: Seq[Feedback],
                          maxModelSize: Int,
                          minModelSize: Int,
                          planningStepsCounter: Int,
-                         kappaPlus: Double,
+                         kappaPlus: INDArray,
                          tolerance: Option[INDArray]) extends Agent with LazyLogging {
   /**
    * Returns the new agent and the chosen action.
@@ -85,9 +86,9 @@ case class ExpSarsaAgent(net: MultiLayerNetwork,
    * @param random      the random generator
    */
   override def chooseAction(observation: Observation, random: Random): Action = {
-    val q0 = Utils.indexed(q(observation), observation.actions)
-    val pr = Utils.egreedy(q0, epsilon)
-    val action = Utils.randomInt(pr)(random)
+    val q0 = q(observation)
+    val pr = egreedy(q0, epsilon)
+    val action = randomInt(pr)(random)
     action
   }
 
@@ -96,7 +97,7 @@ case class ExpSarsaAgent(net: MultiLayerNetwork,
    *
    * @param observation the observation
    */
-  def q(observation: Observation): Policy = net.output(observation.signals)
+  def q(observation: Observation): INDArray = net.output(observation.signals)
 
   /**
    * Returns the fit agent by optimizing its strategy policy and the score
@@ -104,7 +105,7 @@ case class ExpSarsaAgent(net: MultiLayerNetwork,
    * @param feedback the feedback from the last step
    * @param random   the random generator
    */
-  override def fit(feedback: Feedback, random: Random): (Agent, Double) = {
+  override def fit(feedback: Feedback, random: Random): (Agent, INDArray) = {
     val newModel = learnModel(feedback)
     val newNet = net.clone()
     val (newAvg, score) = train(newNet, feedback, avgReward)
@@ -120,47 +121,21 @@ case class ExpSarsaAgent(net: MultiLayerNetwork,
    *
    * @param feedback the feedback from the last step
    */
-  override def score(feedback: Feedback): Double =
-    createData(feedback, avgReward)._4
+  override def score(feedback: Feedback): INDArray = createData(feedback, avgReward)._4
 
   /**
-   * Returns the 3-upla with data for fit
+   * Returns signals, labels, new average, score
    *
    * @param feedback the feedback
    * @param avg      the average rewards
    * @return the input signals, the output label, the new advantage reward, the score
    */
-  private def createData(feedback: Feedback, avg: Double): (INDArray, INDArray, Double, Double) = feedback match {
+  private def createData(feedback: Feedback, avg: INDArray): (INDArray, INDArray, INDArray, INDArray) = feedback match {
     case Feedback(obs0, action, reward, obs1) =>
       // Computes state values
       val q0 = q(obs0)
-      val v0 = Utils.v(q0, action)
-
-      val q1 = Utils.indexed(q(obs1), obs1.actions)
-      val v1 = Utils.vExp(q1, Utils.egreedy(q1, epsilon))
-
-      // Compute new q0 = v1 - Rm + R and delta = v1 - Rm + R - v0
-      val newv0 = v1 + reward - avg
-      val delta = newv0 - v0
-      val score = delta * delta
-
-      // Update average rewards
-      val newAvg = avg + delta * beta
-
-      // Computes labels
-      val labels = q0.dup()
-      labels.putScalar(action, newv0)
-      //            logger.debug("---------------------------------------------------------------")
-      //            logger.debug("  s0     = {}", obs0.signals)
-      //            logger.debug("  action = {}", action)
-      //            logger.debug("  reward = {}", reward)
-      //            logger.debug("  s1     = {}", obs1.signals)
-      //            logger.debug("  q0     = {} -> q0'     = {}", q0, labels)
-      //            logger.debug("  q1     = {}", q1)
-      //            logger.debug("  v0     = {} -> v0'     = {}", v0, newv0)
-      //            logger.debug("  v1     = {}", v1)
-      //            logger.debug("  avg(R) = {} -> avg'(R) = {}", avg, newAvg)
-      //            logger.debug("  delta  = {}", delta)
+      val q1 = q(obs1)
+      val (labels, newAvg, score) = ExpSarsaAgent.createData(q0, q1, action, reward, beta, epsilon, avg)
       (obs0.signals, labels, newAvg, score)
   }
 
@@ -219,17 +194,17 @@ case class ExpSarsaAgent(net: MultiLayerNetwork,
    */
   private def plan(net: MultiLayerNetwork,
                    model: Seq[Feedback],
-                   time: Double,
-                   avg: Double,
-                   random: Random): Double = {
+                   time: INDArray,
+                   avg: INDArray,
+                   random: Random): INDArray = {
     if (model.size > minModelSize) {
       var avg1 = avg
       for {_ <- 1 to planningStepsCounter} {
         val idx = random.nextInt(model.size)
         val feedback = model(idx)
         // compute reward bonus for late transitions
-        val dt = time - feedback.s1.time
-        val bouns = kappaPlus * Math.sqrt(dt)
+        //val dt = time.sub(feedback.s1.time)
+        //val bouns = sqrt(dt).muli(kappaPlus)
         // Create inputs
         val (newAvg, _) = train(net, feedback, avg1)
         avg1 = newAvg
@@ -247,7 +222,7 @@ case class ExpSarsaAgent(net: MultiLayerNetwork,
    * @param feedback the feedback
    * @param avg      the average reward
    */
-  private def train(net: MultiLayerNetwork, feedback: Feedback, avg: Double): (Double, Double) = {
+  private def train(net: MultiLayerNetwork, feedback: Feedback, avg: INDArray): (INDArray, INDArray) = {
     val (inputs, labels, newAvg, score) = createData(feedback, avg)
     // Train network
     net.fit(inputs, labels)
@@ -256,7 +231,7 @@ case class ExpSarsaAgent(net: MultiLayerNetwork,
 }
 
 /** The factory of [[ExpSarsaAgent]] */
-object ExpSarsaAgent {
+object ExpSarsaAgent extends LazyLogging {
 
   /**
    * Returns the Dyna+Agent
@@ -267,18 +242,20 @@ object ExpSarsaAgent {
    */
   def fromJson(conf: ACursor)(noInputs: Int, noActions: Int): ExpSarsaAgent = {
     val tolerance = loadTolerance(conf.downField("tolerance"))
-    val avgReward = conf.get[Double]("avgReward").right.get
+    val avgReward = conf.get[Double]("avgReward").toTry.map(
+      Nd4j.ones(1).muli(_)
+    ).get
     val net = AgentNetworkBuilder.fromJson(conf.downField("network"), "network.zip")(noInputs, noActions, Activation.IDENTITY)
     ExpSarsaAgent(net = net,
       noActions = noActions,
       model = Seq(),
-      beta = conf.get[Double]("beta").toTry.get,
+      beta = conf.get[Double]("beta").toTry.map(Nd4j.ones(1).muli(_)).get,
       avgReward = avgReward,
       maxModelSize = conf.get[Int]("maxModelSize").toTry.get,
       minModelSize = conf.get[Int]("minModelSize").toTry.get,
-      epsilon = conf.get[Double]("epsilon").toTry.get,
-      kappa = conf.get[Double]("kappa").toTry.get,
-      kappaPlus = conf.get[Double]("kappaPlus").toTry.get,
+      epsilon = conf.get[Double]("epsilon").toTry.map(Nd4j.ones(1).muli(_)).get,
+      kappa = conf.get[Double]("kappa").toTry.map(Nd4j.ones(1).muli(_)).get,
+      kappaPlus = conf.get[Double]("kappaPlus").toTry.map(Nd4j.ones(1).muli(_)).get,
       planningStepsCounter = conf.get[Int]("planningStepsCounter").toTry.get,
       tolerance = tolerance)
   }
@@ -318,4 +295,54 @@ object ExpSarsaAgent {
     require(interval._2 >= interval._1)
   }
 
+  /**
+   * Returns output labels, new average, score
+   *
+   * @param q0      the initial action value
+   * @param q1      the final action value
+   * @param action  the action
+   * @param reward  the reward
+   * @param beta    the beta parameter
+   * @param epsilon the epsilon parameter
+   * @param avg     the average reward
+   */
+  def createData(q0: INDArray,
+                 q1: INDArray,
+                 action: Action,
+                 reward: INDArray,
+                 beta: INDArray,
+                 epsilon: INDArray,
+                 avg: INDArray): (INDArray, INDArray, INDArray) = {
+    // Computes state values
+    val v0 = Utils.v(q0, action)
+    val v1 = vExp(q1, egreedy(q1, epsilon))
+
+    // Compute new v0' = v1 - Rm + R and delta = v1 - Rm + R - v0
+    val newv0 = v1.add(reward).subi(avg)
+    val delta = newv0.sub(v0)
+
+    // Update average rewards
+    val newAvg = reward.mul(beta).addi(beta.sub(1).negi().muli(avg))
+    //val newAvg = delta.mul(beta).add(avg)
+
+    // Computes labels
+    val labels = q0.dup()
+    labels.put(action, newv0)
+    logger.whenDebugEnabled {
+      logger.debug("---------------------------------------------------------------")
+      logger.debug("  action  = {}", action)
+      logger.debug("  reward  = {}", reward)
+      logger.debug("  avg(R)  = {}", avg)
+      logger.debug("  q0      = {}", q0)
+      logger.debug("  q1      = {}", q1)
+      logger.debug("  v0      = {}", v0)
+      logger.debug("  v1      = {}", v1)
+      logger.debug("  v0'     = {}", newv0)
+      logger.debug("  delta   = {}", delta)
+      logger.debug("  q0'     = {}", labels)
+      logger.debug("  avg'(R) = {}", newAvg)
+    }
+    val score = pow(delta, 2)
+    (labels, newAvg, score)
+  }
 }